+++ src/main/kotlin/ar/edu/unq/desapp/grupoc/backenddesappapi/webservice/dto/ActiveIntentionResponseDTO.kt
@@ -2,6 +2,7 @@ package ar.edu.unq.desapp.grupoc.backenddesappapi.webservice.dto
 
 import ar.edu.unq.desapp.grupoc.backenddesappapi.model.enums.SYMBOL
 import com.fasterxml.jackson.annotation.JsonProperty
+import java.math.BigDecimal
 import java.time.LocalDateTime
 
 data class ActiveIntentionResponseDTO(
@@ -12,11 +13,11 @@ data class ActiveIntentionResponseDTO(
     val symbol: SYMBOL,
 
     @JsonProperty("nominalAmount")
-    val nominalAmount: Double,
+    val nominalAmount: BigDecimal,
 
     @JsonProperty("cryptoQuote")
-    val cryptoQuote: Double,
+    val cryptoQuote: BigDecimal,
 
     @JsonProperty("localAmount")
-    val localAmount: Double,
+    val localAmount: BigDecimal,
 )
