+++ src/main/kotlin/ar/edu/unq/desapp/grupoc/backenddesappapi/webservice/dto/ExpressIntentionResponseDTO.kt
@@ -3,19 +3,20 @@ package ar.edu.unq.desapp.grupoc.backenddesappapi.webservice.dto
 import ar.edu.unq.desapp.grupoc.backenddesappapi.model.enums.OPERATION
 import ar.edu.unq.desapp.grupoc.backenddesappapi.model.enums.SYMBOL
 import com.fasterxml.jackson.annotation.JsonProperty
+import java.math.BigDecimal
 
 data class ExpressIntentionResponseDTO(
     @JsonProperty("crypto")
     var cryptoAsset: SYMBOL?,
 
     @JsonProperty("nominal_amount")
-    var nominalAmount: Double?,
+    var nominalAmount: BigDecimal?,
 
     @JsonProperty("crypto_quote")
-    var cryptoQuote: Double?,
+    var cryptoQuote: BigDecimal?,
 
     @JsonProperty("operation_amount_ars")
-    var operationAmountARS: Double?,
+    var operationAmountARS: BigDecimal?,
 
     @JsonProperty("first_name")
     var firstName: String?,
